+++ src/main/java/com/evo/NEAT/NodeGene.kt
@@ -4,7 +4,7 @@ package com.evo.NEAT
  * NodeGene represents the nodes of the neural network
  * Created by vishnughosh on 28/02/17.
  */
-class NodeGene(var value: Float) {
+class NodeGene(var value: Double) {
     var incomingCon = ArrayList<ConnectionGene>()
     override fun toString(): String {
         return "NodeGene(value=$value, incomingCon=$incomingCon)"
